requirements.txt
python-twitter==3.4.1

kafka-python==1.4.2
elasticsearch==6.2.0
elasticsearch-dsl==6.1.0
SQLAlchemy==1.2.5
psycopg2-binary==2.7.4

Flask==0.12.2
Flask-SQLAlchemy==2.3.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
future==0.16.0
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
kazoo==2.4.0
MarkupSafe==1.0
oauthlib==2.0.7
python-dateutil==2.7.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tabulate==0.8.2
urllib3==1.22
Werkzeug==0.14.1
